Corrections to datasets and harm_data_numerical data definitions

+++ p2f_api/data/harm_data_numerical.py
@@ -2,8 +2,8 @@
 from ..apilogs import logger
 from .p2f_decbase import baseSQL
 # Third Party Libraries
-from sqlalchemy import BigInteger
-from sqlalchemy import Double
+from sqlalchemy import BigInteger, Integer
+from sqlalchemy import Double, Float
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy import ForeignKey
 # Batteries included libraries
@@ -12,10 +12,10 @@ class harmonized_int_confidence(baseSQL):
     __tablename__ = "p2f_harm_integer_confidence"
     pk_harm_num: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     value: Mapped[int] = mapped_column(BigInteger)
-    upper_conf_interval: Mapped[float] = mapped_column(float)
-    lower_conf_interval: Mapped[float] = mapped_column(float)
-    upper_conf_value: Mapped[int] = mapped_column(int)
-    lower_conf_value: Mapped[int] = mapped_column(int)
+    upper_conf_interval: Mapped[float] = mapped_column(Float)
+    lower_conf_interval: Mapped[float] = mapped_column(Float)
+    upper_conf_value: Mapped[int] = mapped_column(BigInteger)
+    lower_conf_value: Mapped[int] = mapped_column(BigInteger)
     fk_data_record: Mapped[str] = mapped_column(ForeignKey("p2f_harm_data_record.record_hash"))
     fk_data_type: Mapped[int] = mapped_column(ForeignKey("p2f_harm_data_types.pk_harm_data_type"))
 
@@ -30,8 +30,8 @@ class harmonized_float_confidence(baseSQL):
     __tablename__ = "p2f_harm_float_confidence"
     pk_harm_num: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     value: Mapped[float] = mapped_column(Double)
-    upper_conf_interval: Mapped[float] = mapped_column(float)
-    lower_conf_interval: Mapped[float] = mapped_column(float)
+    upper_conf_interval: Mapped[float] = mapped_column(Float)
+    lower_conf_interval: Mapped[float] = mapped_column(Float)
     upper_conf_value: Mapped[float] = mapped_column(Double)
     lower_conf_value: Mapped[float] = mapped_column(Double)
     fk_data_record: Mapped[str] = mapped_column(ForeignKey("p2f_harm_data_record.record_hash"))
